DB implementation
le funzioni per le gestioni dei DB funzionano, rimangono i testing dei server e la popolazione del server main con chiavi ecc...

+++ Progetto_IML/DecBinToPrime.m
@@ -25,8 +25,7 @@ function return_value=DecBinToPrime(img)
         
    end
 
-   return_value=nthprime(bin2dec(temp));
-
+   return_value=int32(nthprime(bin2dec(temp)));
 
 end
